--- lib/colors.rb
@@ -1,1 +1,0 @@
-require_relative "colors/matcher"
